+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+- Rename the public GitHub repository skills command from `skill browse` to `skill search`.
+
 ## [0.2.0] - 2026-04-24
 
 ### Added